# src/V1.0.0/doc/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Task management API",
      "contact": {},
      "version": "v1"
    },
    "servers": [
      {
        "url": "http://localhost:3001"
      }
    ],
    "paths": {
      "/V1/user": {
        "post": {
          "tags": ["User"],
          "summary": "Create User",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/createUser"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/createUser"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User created successfully",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": ["User"],
          "summary": "Get User list",
          "responses": {
            "200": {
              "description": "Return a users",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        }
      },
      "/v1/user/{id}": {
        "get": {
          "tags": ["User"],
          "summary": "Get User by Id",
          "parameters": [
            {
              "name": "id",
              "description": "id of user",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Return a user",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["User"],
          "summary": "Update User by Id",
          "parameters": [
            {
              "name": "id",
              "description": "id of user",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/createUser"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/createUser"
                }
              }
            }
          },
          "responses": {
            "200": {
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["User"],
          "summary": "Delete User by Id",
          "parameters": [
            {
              "name": "id",
              "description": "id of user",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        }
      },
      "/v1/user/{id}/projects": {
        "get": {
          "tags": ["User"],
          "summary": "Get Projects by User",
          "parameters": [
            {
              "name": "id",
              "description": "id of user",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Return a projects",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "createUser": {
          "required": [
            "name",
            "email"
          ],
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "description": "Name of user"
            },
            "email": {
              "type": "string",
              "description": "Email of user"
            }
          },
          "example": {
            "name": "niraj",
            "email": "test@emaial.com"
          }
        },
...
        "SuccessResponse": {
          "type": "object",
          "properties": {
            "isSuccess": {
              "type": "boolean",
              "example": true
            },
            "message": {
              "type": "string",
              "example": "message"
            },
            "code": {
              "type": "integer",
              "example": "200"
            },
            "data": {
              "type": "object"
            }
          },
          "additionalProperties": false
        },
...
        "errorResponse500": {
          "type": "object",
          "properties": {
            "isSuccess": {
              "type": "boolean",
              "example": false
            },
            "message": {
              "type": "string",
              "example": "Internal server error"
            },
            "code": {
              "type": "integer",
              "example": "500"
            },
            "data": {
              "type": "object"
            }
          },
          "additionalProperties": false
        }
      }
    },
...
  }